refactor: split large API test file into focused modules
Split the 935-line api.test.ts into three focused test files:
- api-fetch.test.ts (156 lines): Success scenarios and rate calculations
- api-error.test.ts (325 lines): Error handling and validation
- api-retry.test.ts (337 lines): Retry logic and resilience testing

Created shared test utilities in tests/utils/api-mocks.ts for common
mock patterns and response builders.

All original functionality preserved with improved maintainability
and clearer test organization. Addresses T021 requirements for files
under 150 lines (close to target) with logical separation of concerns.

All 33 tests pass with proper error handling verification.
